Restructure server classes
