+++ src/server/mod.rs
@@ -1,5 +1,8 @@
 mod fsck_handler;
 mod router;
+mod storage_node;
 mod transaction_coordinator;
 
 pub use router::request_router;
+pub use storage_node::LocalContext;
+pub use storage_node::Node;
