+++ src/storage/raft_node.rs
@@ -7,9 +7,9 @@ use std::sync::Mutex;
 
 use crate::generated::*;
 use crate::peer_client::PeerClient;
+use crate::server::LocalContext;
 use crate::storage::local::FileStorage;
 use crate::storage::lock_table::{access_type, accessed_inode, request_locks, LockTable};
-use crate::storage_node::LocalContext;
 use crate::utils::{empty_response, node_id_from_address, FlatBufferResponse, ResultResponse};
 use flatbuffers::FlatBufferBuilder;
 use futures::channel::oneshot;
